src/requirements.txt
Django>=4.1.0,<4.2
django-allauth
django-compressor
gunicorn
markdown
python-decouple
django-htmx
django-storages
black
boto
boto3 
psycopg2
django-hosts
pillow
django-admin-rangefilter
markdown
django-crispy-forms
crispy-tailwind
pyyaml